apps/backend/requirements.txt
# =============================================================================
# Unified Requirements for All Services
# =============================================================================

# -----------------------------------------------------------------------------
# Django Backend
# -----------------------------------------------------------------------------
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.9.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.0
gunicorn>=21.0

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
psycopg2-binary>=2.9
pgvector>=0.2
neo4j==6.0.3
redis>=5.0
elasticsearch>=8.0

# -----------------------------------------------------------------------------
# LangChain & RAG
# -----------------------------------------------------------------------------
langgraph==0.2.61
langchain==0.3.26
langchain-openai==0.3.8
langchain-community==0.3.27
openai>=1.0

# -----------------------------------------------------------------------------
# API Framework
# -----------------------------------------------------------------------------
uvicorn==0.38.0
jinja2>=3.1
fastapi

# -----------------------------------------------------------------------------
# ML & Recommendation
# -----------------------------------------------------------------------------
torch>=2.1
scikit-learn==1.7.2
pandas==2.3.3
numpy==2.3.5

# -----------------------------------------------------------------------------
# Data Science & Analytics
# -----------------------------------------------------------------------------
jupyter==1.1.1
matplotlib==3.10.7
seaborn==0.13.2
openpyxl==3.1.5
xlrd==2.0.2

# # -----------------------------------------------------------------------------
# # Web Crawling
# # -----------------------------------------------------------------------------
# playwright
# selenium
# beautifulsoup4
# global-land-mask

# -----------------------------------------------------------------------------
# Authentication
# -----------------------------------------------------------------------------
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest>=7.0
hypothesis>=6.0

# -----------------------------------------------------------------------------
# Logging
# -----------------------------------------------------------------------------
python-json-logger>=2.0

# -----------------------------------------------------------------------------
# Utils
# -----------------------------------------------------------------------------
python-dotenv==1.1.1
pydantic==2.12.3
Pillow>=10.0.0
requests==2.32.5
